Harden number parsing and error surfacing
- Parse numbers with the invariant culture in both the validator and
  the number handler, so comma-decimal locales no longer misparse a
  '.' decimal. Reject NaN/Infinity literals (e.g. overflowing values)
  instead of accepting them silently.
- Guard the terminal pop in the converter: an empty operand stack now
  throws a domain MissingOperandException rather than leaking a BCL
  "Stack empty" InvalidOperationException (e.g. for "()").
- Console app: add an 'exit' command, guard against null input on EOF
  so it no longer spins forever, and fix the "occured" typo.
