+++ Calculator.BL/Implementations/TokenActionHandlers/NumberTokenHandler.cs
@@ -1,6 +1,8 @@
 ﻿using Calculator.BL.Abstractions;
 using Calculator.Core.Abstractions;
 using Calculator.Core.Implementations;
+using Calculator.Kernel;
+using System.Globalization;
 
 namespace Calculator.BL.Implementations.TokenActionHandlers
 {
@@ -8,7 +10,7 @@ namespace Calculator.BL.Implementations.TokenActionHandlers
     {
         public void HandleAction(Token token, Stack<ICalculatable> operands)
         {
-            float number = float.Parse(token.Value);
+            float number = float.Parse(token.Value, NumberStyles.Float, CultureInfo.InvariantCulture);
             operands.Push(new CalculatableNumber(number));
         }
     }
